Extract common methods from events_processor and transactions_processor
